Load / Save method implementations
Load pulls the data from etcd and explodes it into a full data object
Save flattens the data object and stores int etcd at the corresponding keys.
Does a best effort to mirror changes in directories and values.

If setting a value at a previous directory, remove the directory and set
a value

If setting a nested value at a previous value, delete the value and set
the approriate keys